started the new uart_reader V5 and fixed the database gui
