+++ mcu/src/mcu/temp_storage_of_scripts/databaseUtils.py
